pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "techain-ia"
version = "0.1.0"
description = "Production-grade algorithmic trading engine"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "prometheus-client>=0.20",
    "aiohttp>=3.9",
    "asyncio-mqtt>=0.16",
    "pydantic>=2.6",
    "scipy>=1.12",
    "python-dateutil>=2.9",
    "rich>=13.7",
    "typer>=0.12",
    "python-binance>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
brokers = [
    "alpaca-py>=0.26",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
